Сделал отображение фотографии пользователя на фронтенде.

+++ src/main/java/ru/skypro/homework/repository/UserRepo.java
@@ -5,6 +5,6 @@ import ru.skypro.homework.entity.UserEntity;
 
 import java.util.Optional;
 
-public interface UserRepo extends JpaRepository<UserEntity, Integer> {
+public interface UserRepo extends JpaRepository<UserEntity, Long> {
     Optional<UserEntity> findByEmail(String username);
 }
